refactor orchestrator structure
