+++ tools/pipeline_perf_test/orchestrator/lib/process/__init__.py
@@ -1,0 +1,5 @@
+from .stats import ProcessStats
+
+__all__ = [
+    "ProcessStats",
+]
