+++ tools/pipeline_perf_test/orchestrator/lib/process/deployed_process/__init__.py
@@ -1,0 +1,9 @@
+from .base import DeployedProcess
+from .docker import DockerProcess
+from .kubernetes import K8sDeployedResource
+
+__all__ = [
+    "DeployedProcess",
+    "DockerProcess",
+    "K8sDeployedResource",
+]
